Extract Linux TCP reactor module
Move TCP-specific reactor state, lifecycle, I/O helpers, and focused tests into dedicated socket/tcp modules while preserving shared TCP/UDP dispatch in socket.rs.


Copilot-Session: b5a1a347-37a8-4246-8bbc-306590921475
